self_infer/requirements.txt
gdown
opencv-python
dlib
Pillow
dotmap
imutils
matplotlib
PyYAML
tensorflow_addons
keras_cv_attention_models
pandas
scikit-learn